requirements.txt
fastapi==0.115.7
uvicorn
pandas==2.2.3
openai
python-dotenv
python-multipart